Use the same service for statefulSet and deployment modes
There is no reason to use an headless service in that case. The
statefulSet resource is used here to maintain access to the same volume
even if the workload is redeployed.

As the statefulSet is always scaled to 1, the service is used to access
the workload and not to do an auto discovery of each replicas.

So:

- The headless service has been removed
- The service section in the statefulSet section (in values.yaml) has
  been removed
- The service resource is now used for both workload modes