charts/vmalert: fix format of flags when using notifiers (#536)
Previously, it was creating a list entry only for notifiers which had some specific options provided. That created a list of items where only non-empty options were specified, so that would be applied only to the first notifier.
Fix changes templating to create empty entries for items where option is not enabled, so that each notifier will only receive its own option respectively.