feat(crd): view and admin aggregate roles for CRDs
Kubernetes has view and admin roles that aggregate permissions from
other clusterroles that have the right annotations.

The advantage to this is that the view clusterrole doesn't need to be
edited whenever a new CRD is installed, but can still be extended so
view and admin also provide view and admin access to new CRDs.

The functionality is similar to drop-in files in systemd and other
linux/unix tooling.

The change adds the aggregate role functionality to the victoria-metrics-operator
helm chart. So whenever the CRDs are installed the user can also choose
to install the clusterroles that extend the view and admin default
clusterroles.