Validate manifests with `kubectl` in CI (#852)
* 👷 Validate with kubectl before install

Would be nice to have it in the linting CI, however this is the only place in CI that we have a cluster to work with. In order to validate the manifests properly, `--dry-run=server` is required

* 💚 Validate manifests after install

Since we might rely on some CRDs we need to validate post-install. Also since we only test one chart per CI run, and setup k3s per job, then we can use `--skip-clean-up` to keep the deployment around

* ⬆️ Bump ct to use new --skip-clean-up flag

* 🩹 Use same namespace for dry-run

No issues with using `default`, it's mostly for consistency

* 🐛 Validate is required to detect capabilities

Without it, the incorrect API versions will not be installed. i.e. `if .Capabilities.APIVersions.Has "..."` will be incorrect

* 💚 We want to validate the manifests generated

So replace is fine here

* ci: validate manifests before running ct testing

* ci: do not skip crds

* ci: skip run for k8s-stack chart due to helm validation issue

---------

Co-authored-by: Zakhar Bessarab <z.bessarab@victoriametrics.com>