add support for globally configurable `timeField` and `msgField` parameters (#2572)
These changes are required to prepare the migration of the underlying implementation to an application that uses the `/internal/insert` handler, which does not support the `VL-Time-Field`, `VL-Msg-Field` headers. Therefore, the client must handle this logic on its own side.

Since it is challenging and unnecessary to use different _msg and _time fields for each remoteWrite, the rw.msgField parameter is now deprecated - clients should use the global msgField instead. Existing configurations will not be affected, as the new logic takes the old parameter into account.